appliance: Disable LD_PRELOAD libSegFault in virt-rescue shell.
Commit c598e14052399f19d36d7874d7133cba779bd3c2 exports
LD_PRELOAD=/lib64/libSegFault.so so that guestfsd and processes that
it runs will produce descriptive stack traces.

However it exports it to everything including the virt-rescue shell.
But if we are chrooting into a guest which doesn't have this file (or
indeed any guest) we should not use LD_PRELOAD.

Therefore unset LD_PRELOAD in the rescue shell.

See also related commit 21e5fc811ea3708ef0c768467ffe53368bc435e9.