firstboot: Install both systemd and SysV init scripts.
This should be safe, and avoids the whole issue of whether a guest
uses systemd or not.  Also it is possible to switch init systems
(eg. by specifying init= on the command line) so it's not even clear
if a guest "is" systemd.

Thanks: Tom Gundersen on #systemd.

This updates/fixes commit 596de56a436d56886616f5cc213bb72a4f6f5561.