appliance: Remove cruft and run udevd directly.
This commit removes a lot of cruft from old RHEL which was required to
start udev.  Instead we run the daemon directly.

Also it removes the fallback path which created some device nodes in
/dev by hand.  This likely did not work.  This also removes the
dependency on MAKEDEV which is long gone from most distros.

This is required for Ubuntu, since the /etc/init.d/udev script didn't
actually work.  The most common symptom was that /dev/virtio-ports/*
was not created because udev didn't process the cold start events.