builder: index: Add checksum[sha512]=... option and deprecate template sigs.
We don't need to sign individual templates, provided the signed index
file contains the cryptographically secure checksums of the templates.
Since the index file cannot be interfered with and the checksums
cannot be spoofed, this is sufficient to ensure the integrity of the
templates.

Note the template signatures are now deprecated and will be removed in
libguestfs 1.26.  To avoid any backwards compatibility issues, I will
backport the relevant changes to the stable-1.24 branch.

Thanks Dennis Gilmore for pointing out that we could do this.

(cherry picked from commit 4b73e0be29302da0ae8c5c79a51e770cbe6ed7cf)