builder: Only use virt-resize --no-sparse when writing to a block device.
Previously we would use the virt-resize --no-sparse option if the
final output file was a block device.  This is safe, but unnecessary
for example if virt-resize was used as an intermediate step.  So only
use this option if virt-resize is actually writing to the block
device.

(cherry picked from commit 0037820ee003bfbc71bf35b18fb3a066cbf298f8)