fish/uri: Initialize some variables to avoid a warning.
The warning only appears when we set CFLAGS to be the same as used by
RPM, which you can do by setting:

  CFLAGS="$(rpm --eval '%{__global_cflags}')"

before running ./configure.

FWIW the current value of that RPM expression (in Fedora 20) is:

  -O2 -g -pipe -Wall -Wp,-D_FORTIFY_SOURCE=2 -fexceptions -fstack-protector-strong --param=ssp-buffer-size=4 -grecord-gcc-switches

It is not clear which option precisely enables the warning.

(cherry picked from commit cd9f7411741290add87910d531a9cd8685b3be91)