appliance: Fix 'make clean' in appliance subdirectory.
Commit 71dfd73042db24217329d815bf5eab26f413d97f broke 'make clean' in
the appliance subdirectory, so it would remove all the supermin.d/*
files, but without removing stamp-supermin so you couldn't rebuild
supermin.d/ with a 'make' afterwards.

In any case, 'make clean' is not supposed to delete the appliance
(which is expensive to rebuild).

Fix this so that 'make clean' doesn't remove supermin.d/* files.
'make distclean' can be used to delete this, and since it also removes
stamp-supermin, it will be recreated properly by 'make'.