Backport virt-sysprep from libguestfs 1.24
Copy the virt-sysprep code from libguestfs 1.24.x, bringing all the
features of it to libguestfs 1.20.x. This is done as single commit
because cherry-picking individual commits is simply not worth it, and
would require to cherry-pick a big number of commits, possibly bringing
in new features implemented elsewhere, etc.

Move existing files (where possible) or import them from 1.24.x,
adapting the current ocaml-based tools to the updated functions in
mllib.