virt-p2v: Install the binary in libexecdir.
Previously it wasn't installed at all.  However it makes sense to
install it somewhere so that it can be placed in a package (a
requirement if we're going to create a kickstart).  By placing it in
libexecdir it is hopefully clear that it should not be used.

Note on Red Hat distros, libexecdir is usually /usr/libexec.  On
Debian it is usually /usr/lib/libguestfs.