podwrapper: Put the =encoding line at the correct place in the input (RHBZ#1109174).
Commit c4dc70f8c43a7faae2a8698027a4ed23d316dfa0 broke the man pages
for the Perl scripts in the tools/ directory.

It inserted the =encoding line at the top of the file, instead of in
front of the first =head1 entry, and this meant that the #!/usr/bin/perl
and the prologue became a part of the POD, which was not intended.

(cherry picked from commit 1148bd91ce2c0bebc65f9855368fd026aaf9f603)