v2v: Allow RHEV UUIDs to be specified on the command line.
Currently various UUIDs used by `-o rhev' are generated randomly.
This means they are different on every run of the program.

Allow some of them to be specified on the command line.  (If not
specified they are still generated randomly).

This allows RHEV-M to run virt-v2v several times (eg to examine the
OVF output) while having paths to files and OVF be stable across runs.