v2v: Add display, removable disks, and network interfaces.
These are now collected from the source metadata and (where
possible) passed through to the destination.

For reference, see the metadata structure of the old virt-v2v (roughly
equivalent to Types.source in the new code):

https://git.fedorahosted.org/cgit/virt-v2v.git/tree/metadata-format.txt

As you can see the new code now covers all the metadata features of
the old code.