v2v: source_libvirt: Replace create_xml ?dir parameter hack.
Instead of the ?dir parameter which was essentially a hack, allow
callers of create_xml to provide functions that map
<source file="..."> or <source dev="..."> to qemu URIs.

The default for these mappings is the identity function.

For -i libvirtxml we replace the old absolute_path_of_disk function
with a source file mapping function which does the same thing.

There is no functional change in this patch.