build: isolate the libdb tools search in an own autoconf macro
Create an autoconf macro to deal with the possible combinations of names
and versions of the libdb tools. This also eases the addition of a libdb
version, or a new pattern for the names of tools.

There is no functional change.

(cherry picked from commit a293ac1e911052c52e3110800a5b7b2c5f1bd37d)