resize: fix 'No space left on device' problem when copying to an extended partition (RHBZ#1169015)
Because of the size of an extended partition reported by Linux is always 1024
bytes, so it will always fail to copy to an extended partition.

This patch fixes this problem by copying to the offset of an extended
partition in the destination disk.

Signed-off-by: Hu Tao <hutao@cn.fujitsu.com>
(cherry picked from commit 9d6f0b6a86d68438b27a3d783677c63f39ec6627)