drives: Remove check for ':' in filename (RHBZ#1128942).
Modern qemu can now handle this properly.  ':' is only special if what
precedes it looks like a transport, so:

  qemu-system-x86_64 -drive foo:bar     .. fails
  qemu-system-x86_64 -drive ./foo:bar   .. works

Thus by adding ./ in front of relative paths that contain ':' we can
work around this.

In addition, this broke iscsi:// URIs because iSCSI target names
routinely contain ':' characters.

This is a modified version of upstream
commit a95214b1985e694946e3426120a6fdc13a3f081f, which fixed upstream
RHBZ#811649.