lib: Fix static linking by forcing launch-*.o objects to be always linked.
Because the launch-*.o objects only contain statically scoped
declarations, the linker doesn't need to link them in.  This causes a
problem when static linking as they will be left out of the final
binary, and so no backends will be available.

This is essentially the same problem as described here:

https://stackoverflow.com/questions/1202494/why-doesnt-attribute-constructor-work-in-a-static-library